==> 06_textures_soilgrids_validate.ipynb <==
import numpy as np  # type: ignore
import pandas as pd  # type: ignore

import os

import matplotlib.pyplot as plt  # type: ignore
import subprocess

import rasterio
from rasterio.plot import show
from rasterstats import zonal_stats
from rasterio.warp import calculate_default_transform, reproject, Resampling

%matplotlib inline

path = r"data\EU_SoilHydroGrids_250m_2017Feb08"

grid_cells = {}

files = []
folders = []
# r=root, d=directories, f = files
for r, d, f in os.walk(path):
    for folder in d:
        folders.append(folder)
        
        for r1, d1, f1 in os.walk(os.path.join(path, folder)):
        
            for file in f1:
                if f"FC_M_sl1_{folder}.tif" in file:
                    files.append(os.path.join(r1, file))
                    data_src = rasterio.open(os.path.join(r1, file))
                    srs = data_src.crs
                    bounds = data_src.bounds
                    sub_obj = {
                        'grid_id': folder,
                        'srs': srs,
                        'bounds': bounds
                    }
                    grid_cells.update({ folder: sub_obj})
                    

for folder in folders:
    print(folder)

for file in files:
    print(file)

display(grid_cells)

import fiona
from shapely.geometry import box, Polygon, shape, mapping
from collections import OrderedDict

schema = {
    "geometry": "Polygon",
    "properties": OrderedDict([
        ("grid_id", "str:200"),
    ])
}

def make_poly_from_bounds(bounds):
    # (minx, miny, maxx, maxy) or a Polygon instance
    bbox = box(bounds.left, bounds.bottom, bounds.right, bounds.top)
    return bbox

def make_feature_from_dict(grid_obj):
    return {
        "geometry": mapping(make_poly_from_bounds(grid_obj.get('bounds'))),
        "properties": {"grid_id": grid_obj.get('grid_id')}
    }
    
features = []

for obj in grid_cells.values():
    # display(obj)
    feat = make_feature_from_dict(obj)
    features.append(feat)

crs_wkt = data_src.crs.wkt

# From https://doi.org/10.5281/zenodo.3446747 Grid tiles for identifying EU-SoilHydroGrids tiles
with fiona.open(r"data\EU_SoilHydroGrids_250m_2017Feb08\grid_cells.shp", "w", driver="ESRI Shapefile", schema=schema, crs_wkt=crs_wkt) as collection:
    collection.writerecords(features)
    print(len(collection))
    collection.flush()



path_eesti_250m = r"data\Estonia_EU_Hydrosoilgrids\Estonia_250m"

for r, d, f in os.walk(path_eesti_250m):
    for folder in d:
        for n in range(1,8):
            try:
                # We handle the connections with "with"
                with rasterio.open(os.path.join(r, folder, f"FC_M_sl{n}_{folder}.tif")) as src:
                    FC = src.read(1, masked=True)
                    nodataval = src.nodata

                with rasterio.open(os.path.join(r, folder, f"WP_M_sl{n}_{folder}.tif")) as src:
                    WP = src.read(1, masked=True)

                # Allow division by zero
                np.seterr(divide='ignore', invalid='ignore')

                # Calculate NDVI
                AWC = FC.astype(np.uint8) - WP.astype(np.uint8)
                
                # write_out_awc_raster(folder, n)
                # Define spatial characteristics of output object (basically they are analog to the input)
                kwargs = src.meta

                # Update kwargs (change in data type)
                kwargs.update(dtype=rasterio.uint8, count = 1, nodata = nodataval, masked = True)

                # Let's see what is in there
                print(kwargs)

                with rasterio.open(os.path.join(r, folder, f"AWC_M_sl{n}_{folder}.tif"), 'w', **kwargs) as dst:
                    dst.write_band(1, AWC.astype(rasterio.uint8))
                print(subprocess.check_output("gdalinfo " + os.path.join(r, folder, f"AWC_M_sl{n}_{folder}.tif"), shell=True))
                        
            except Exception as ex:
                print(ex)


## Now loading the soil polygons

- actually, load the big soil shp, only keep some columns and write out as shp again
- then run the rasterstats with keeping the geojson true and orig_fid attribute in particular
- then build a dataframe out of it and start developing the magic for the different layer aggregations
- once the AWC is mean per SOL_N layers and depth based on the original measured AWC EU hydrogrid depths join this AWC back to the main soil db (and basically only overwrite the SOL_AWC_1-4)

import numpy as np  # type: ignore
import pandas as pd  # type: ignore

import fiona  # type: ignore
from fiona.crs import from_epsg # type: ignore
import geopandas as gpd  # type: ignore

eesti_soil_red1_validatesoil = gpd.read_file("../data_deposit/EstSoil-EH_v1.0.shp", encoding='utf-8')


columns_to_keep = [ "orig_fid",
                    "nlayers",
                    "SOL_ZMX",
                    "SOL_Z1",
                    "SOL_Z2",
                    "SOL_Z3",
                    "SOL_Z4",
                   'SOL_CLAY1',
                   'SOL_SILT1',
                   'SOL_SAND1',
                   'SOL_ROCK1',
                    'SOL_BD1',
                   'SOL_SOC1',
                    "geometry"]

eesti_soil_red1_validatesoil_short = eesti_soil_red1_validatesoil[columns_to_keep].copy()
del(eesti_soil_red1_validatesoil)

eesti_soil_red1_validatesoil_short.to_file('../data_deposit/EstSoil-EH_sand_silt_coarse_tmp.shp', encoding='utf-8')



eesti_soil_red1_validatesoil_short.describe()

import numpy as np  # type: ignore
import pandas as pd  # type: ignore

import functools
import statistics
import operator

import fiona  # type: ignore
from fiona.crs import from_epsg # type: ignore
import geopandas as gpd  # type: ignore

import rasterio
from rasterio.plot import show
from rasterstats import zonal_stats
from rasterio.warp import calculate_default_transform, reproject, Resampling

comp_l1 = {
    'sand': 'soil250_grid_sand_sd1_3301.tif',
    'silt': 'soil250_grid_silt_sd1_3301.tif',
    'clay': 'soil250_grid_clay_sd1_3301.tif',
    'rock': 'soil250_grid_coarsefrag_sd1_3301.tif',
    'bd': 'soil250_grid_bulkdens_sd1_3301.tif',
    'soc': 'soil250_grid_soc_sd1_3301.tif'
}

with fiona.open('../data_deposit/EstSoil-EH_sand_silt_coarse_tmp.shp') as vector_src:
    
    src_crs = vector_src.crs
    display(src_crs)
    src_schema = vector_src.schema
    display(src_schema)

    src_schema['properties']["mean"] = "float"
    src_schema['properties']["std"] = "float"
    
    # for layer in range(1,8):
    for layer in comp_l1.keys():
        
        outputs = zonal_stats(vector_src,
                # f"C:\\dev\\05_geodata\\soil\\soilgrids_download\\soil250_grid_sand_sd{layer}_3301.tif",
                f"C:\\dev\\05_geodata\\soil\\soilgrids_download\\{comp_l1[layer]}",
                stats="mean std",
                all_touched=True, geojson_out=True)
    
        with fiona.open(f"../data_deposit/EstSoil-EH_{layer}_zonal_layer.shp", "w", driver="ESRI Shapefile", schema=src_schema, crs=src_crs) as collection:
            collection.writerecords(outputs)
            print(len(collection))
            collection.flush()


# load geodataframes
for layer in range(1,8):
    next_layer = gpd.read_file(f"../data_deposit/EstSoil-EH_sand_silt_coarse_tmp_zonal_layer_{layer}.shp", encoding='utf-8')
    display(next_layer.isnull().sum())
    display(next_layer.sample(10))
    display(next_layer.dtypes)
    fig, ax = plt.subplots()
    fig = next_layer["mean"].hist(ax=ax)
    plt.show()
    fig, ax = plt.subplots()
    fig = next_layer["mean"].plot(ax=ax)
    plt.show()
    fig, ax = plt.subplots()
    fig = next_layer["std"].hist(ax=ax)
    plt.show()
    fig, ax = plt.subplots()
    fig = next_layer["std"].plot(ax=ax)
    plt.show()
    # next_layer["mean"] = next_layer.isnull().apply(lambda x: x["majority"])
    # display(next_layer.isnull().sum())
    # because to save memory space, the factions after comma were stored as Byte / UInt8
    # next_layer['mean'] = next_layer['mean'] / 100.0
    next_layer.to_file(f"../data_deposit/EstSoil-EH_sand_silt_coarse_tmp_zonal_layer_{layer}.shp", encoding='utf-8')
    
# do_layer_avg_1_4

import numpy as np  # type: ignore
import pandas as pd  # type: ignore

import fiona  # type: ignore
from fiona.crs import from_epsg # type: ignore
import geopandas as gpd  # type: ignore

comp_l1 = {
    'sand': 'soil250_grid_sand_sd1_3301.tif',
    'silt': 'soil250_grid_silt_sd1_3301.tif',
    'clay': 'soil250_grid_clay_sd1_3301.tif',
    'rock': 'soil250_grid_coarsefrag_sd1_3301.tif',
    'bd': 'soil250_grid_bulkdens_sd1_3301.tif',
    'soc': 'soil250_grid_soc_sd1_3301.tif'
}


# layer1 = gpd.read_file("data/eesti_soil_red1_fix_geo_awc_zonal_layer_1.shp", encoding='utf-8')
# layer1.drop(columns=['median','std'], inplace=True)
# layer1.rename(columns={"mean" : "AWC_L1"}, inplace=True)

is_first = True
layer1 = eesti_soil_red1_validatesoil_short

# for layer in range(2,8):
for layer in comp_l1.keys():
    next_layer = gpd.read_file(f"../data_deposit/EstSoil-EH_{layer}_zonal_layer.shp", encoding='utf-8')
    next_layer.rename( columns = {
        "mean" : "mean_1_" + str(layer),
        "std" : "std_1_" + str(layer)},
                      inplace=True)
    next_layer.drop( columns = [
                            "nlayers",
                            "SOL_ZMX",
                            "SOL_Z1",
                            "SOL_Z2",
                            "SOL_Z3",
                            "SOL_Z4",
                            "geometry"], inplace=True)
    layer1 = pd.merge(left=layer1, right=next_layer, left_on='orig_fid', right_on='orig_fid', how='left')

display(layer1.sample(10))
display(layer1.dtypes)

eesti_soil_red1_validatesoil = gpd.read_file("../data_deposit/EstSoil-EH_v1.0.shp", encoding='utf-8')


columns_to_keep = ["orig_fid",
                   'SOL_CLAY1',
                   'SOL_SILT1',
                   'SOL_SAND1',
                   'SOL_ROCK1',
                    'SOL_BD1',
                   'SOL_SOC1']

eesti_soil_red1_validatesoil_phys = eesti_soil_red1_validatesoil[columns_to_keep].copy()
del(eesti_soil_red1_validatesoil)

layer1 = pd.merge(left=layer1, right=eesti_soil_red1_validatesoil_phys, left_on='orig_fid', right_on='orig_fid', how='left')

display(layer1.sample(10))
display(layer1.dtypes)

from sklearn.metrics import r2_score, mean_squared_error
import matplotlib.pyplot as plt
%matplotlib inline

import numpy as np
from sklearn import datasets, linear_model
import math

for i in ['SOL_CLAY1',
           'SOL_SILT1',
           'SOL_SAND1',
           'SOL_ROCK1',
            'SOL_BD1',
           'SOL_SOC1']:
    layer1.rename( columns = { i : i.lower()}, inplace=True)

for layer in comp_l1.keys():
    print('----------------------')
    print(layer)
    # display(layer1[f"sol_{layer}1"].isnull().sum())
    # display(layer1["mean_1_" + str(layer)].isnull().sum())
    layerT = layer1.dropna()
    layerT["diff_1_" + str(layer)] = layerT[f"sol_{layer}1"] - layerT["mean_1_" + str(layer)]
    layerT["abs_diff_1_" + str(layer)] = layerT["diff_1_" + str(layer)].apply(abs)
    display(layerT["abs_diff_1_" + str(layer)].describe())
    display(layerT["std_1_" + str(layer)].describe())
    # Calculation of Mean Squared Error (MSE) 
    # RMSE = mean_squared_error(layerT[f"sol_{layer}1"].to_numpy(), layerT["mean_1_" + str(layer)].to_numpy(), squared=False)
    MSE = mean_squared_error(layerT[f"sol_{layer}1"].to_numpy(), layerT["mean_1_" + str(layer)].to_numpy())
    r2 = r2_score( layerT[f"sol_{layer}1"].to_numpy(), layerT["mean_1_" + str(layer)].to_numpy() )
    print(f"RMSE: {math.sqrt(MSE)}  MSE: {MSE}   R2:{r2}")
    print(layer)
    print('#################')
    

# mean_1_sand     float64
# std_1_sand      float64
# SOL_SAND1

# mean_1_silt     float64
# std_1_silt      float64
# SOL_SILT1

# mean_1_clay     float64
# std_1_clay      float64
# SOL_CLAY1

# mean_1_rock     float64
# std_1_rock      float64
# SOL_ROCK1 

# mean_1_bd       float64
# std_1_bd        float64
# SOL_BD1

# mean_1_soc      float64
# std_1_soc       float64
# SOL_SOC1  




layerT["abs_diff_1_sand"].describe()

layer_B = layerT[layerT['sol_sand1'] < 100]
layer_B["abs_diff_1_sand"].describe()

display(layer1.isnull().sum())

display(layer1.loc[layer1['AWC_L7'].isnull()])

cache_row = {}

for idx, row in layer1.loc[layer1['AWC_L7'].isnull()].iterrows():
    
    display(f"found {idx}, take {idx-2}")
    
    layer1.loc[idx,'AWC_L1'] = layer1.loc[idx-2,'AWC_L1']
    layer1.loc[idx,'AWC_L2'] = layer1.loc[idx-2,'AWC_L2']
    layer1.loc[idx,'AWC_L3'] = layer1.loc[idx-2,'AWC_L3']
    layer1.loc[idx,'AWC_L4'] = layer1.loc[idx-2,'AWC_L4']
    layer1.loc[idx,'AWC_L5'] = layer1.loc[idx-2,'AWC_L5']
    layer1.loc[idx,'AWC_L6'] = layer1.loc[idx-2,'AWC_L6']
    layer1.loc[idx,'AWC_L7'] = layer1.loc[idx-2,'AWC_L7']
    
    display(f"found {row['AWC_L1']}, put {layer1.loc[idx,'AWC_L7']}")

layer1.loc[427,'AWC_L1']

import functools
import statistics
import operator

known_depths_to_layer = [  
    (0,1),
    (50,2),
    (150,3),
    (300,4),
    (600,5),
    (1000,6),
    (2000,7)
]

def get_aggregate_awc_for_depths(layer_top, layer_bottom, SOL_AWC1, awc_gradient_values, known_depths_to_layer):
    filt = list(filter(lambda x: x[0] >= layer_top and x[0] <= layer_bottom, known_depths_to_layer))
    lays = list(map(lambda x: x[1], filt))
    
    if len(lays) <= 0:
        position = statistics.mean([layer_top, layer_bottom])
        if position > known_depths_to_layer[6][0]:
            # return [7]
            lays = [7]
        elif position < known_depths_to_layer[0][0]:
            # return [1]
            lays = [1]
        else:
            diffs = []
            for i in range(1,8):
                diffs.append( (statistics.stdev([position, known_depths_to_layer[i-1][0]]), i) )
            diffs.sort(key = operator.itemgetter(0))
            lays = [ diffs[0][1] ]

    vals = statistics.mean(list(map(lambda x: awc_gradient_values[int(x)], lays)))
    return vals


def aggregate_over_depths(row):
    
    SOL_AWC1 = row['SOL_AWC1']
    SOL_AWC2 = row['SOL_AWC2']
    SOL_AWC3 = row['SOL_AWC3']
    SOL_AWC4 = row['SOL_AWC4']
    
    nlayers = row['nlayers']
    SOL_ZMX = row['SOL_ZMX']
    SOL_Z1 = row['SOL_Z1']
    SOL_Z2 = row['SOL_Z2']
    SOL_Z3 = row['SOL_Z3']
    SOL_Z4 = row['SOL_Z4']
    
    SOL_Z2_new = row['SOL_Z2']
    SOL_Z3_new = row['SOL_Z3']
    SOL_Z4_new = row['SOL_Z4']
    
    AWC_L1 = row['AWC_L1']
    AWC_L2 = row['AWC_L2']
    AWC_L3 = row['AWC_L3']
    AWC_L4 = row['AWC_L4']
    AWC_L5 = row['AWC_L5']
    AWC_L6 = row['AWC_L6']
    AWC_L7 = row['AWC_L7']
    
    awc_gradient_values_pre = np.array([AWC_L1, AWC_L2, AWC_L3, AWC_L4, AWC_L5, AWC_L5, AWC_L6, AWC_L7])
    awc_gradient_values = awc_gradient_values_pre / 100
    
    if nlayers >= 1:
        # depth from top to bottom of 1st layer is exactly SOL_Z1
        layer_top = 0
        layer_bottom = SOL_Z1
        SOL_AWC1 = get_aggregate_awc_for_depths(layer_top, layer_bottom, SOL_AWC1, awc_gradient_values, known_depths_to_layer)
    
    if nlayers >= 2:
        layer_top = SOL_Z1
        layer_bottom = SOL_Z1 + SOL_Z2
        SOL_AWC2 = get_aggregate_awc_for_depths(layer_top, layer_bottom, SOL_AWC2, awc_gradient_values, known_depths_to_layer)
        SOL_Z2_new = layer_bottom
    
    if nlayers >= 3:
        layer_top = SOL_Z1 + SOL_Z2
        layer_bottom = SOL_Z1 + SOL_Z2 + SOL_Z3
        SOL_AWC3 = get_aggregate_awc_for_depths(layer_top, layer_bottom, SOL_AWC3, awc_gradient_values, known_depths_to_layer)
        SOL_Z3_new = layer_bottom
    
    if nlayers >= 4:
        layer_top = SOL_Z1 + SOL_Z2 + SOL_Z3
        layer_bottom = SOL_Z1 + SOL_Z2 + SOL_Z3 + SOL_Z4
        SOL_AWC4 = get_aggregate_awc_for_depths(layer_top, layer_bottom, SOL_AWC4, awc_gradient_values, known_depths_to_layer)
        SOL_Z4_new = layer_bottom
    
    return pd.Series([SOL_Z1, SOL_Z2_new, SOL_Z3_new, SOL_Z4_new, SOL_AWC1, SOL_AWC2, SOL_AWC3, SOL_AWC4])

layer1[['SOL_Z1', 'SOL_Z2', 'SOL_Z3', 'SOL_Z4',
        'SOL_AWC1', 'SOL_AWC2', 'SOL_AWC3', 'SOL_AWC4']] = layer1.apply(lambda x: aggregate_over_depths(x), axis=1)

display(layer1.sample(10))

layer1.to_file(f"data/eesti_soil_red1_texture_fix_geo_redo_awc_merged_layers.shp", encoding='utf-8')


display(layer1.isnull().sum())

import math

math.sqrt(4)
